Add PickerBase base class
