+++ bokehjs/src/lib/models/widgets/index.ts
@@ -6,6 +6,8 @@ export {CheckboxGroup}       from "./checkbox_group"
 export {Checkbox}            from "./checkbox"
 export {ColorPicker}         from "./color_picker"
 export {DatePicker}          from "./date_picker"
+export {TimePicker}          from "./time_picker"
+export {DatetimePicker}      from "./datetime_picker"
 export {DateRangeSlider}     from "./date_range_slider"
 export {DateSlider}          from "./date_slider"
 export {DatetimeRangeSlider} from "./datetime_range_slider"
